Updated how we find the .git directory in bootstrap.
When building from git, bootstrap will (among other things) install
pre-commit and post-rewrite hooks into the .git/hooks directory of the
mesos tree. However the current implementation always assumes that .git
exists in the same directory as the bootstrap file. This may not always
be true.

Most notably, it is not true if the mesos tree is included as a
submodule inside another project. When included as a submodule, .git is
no longer a directory, but rather a file whose text contains a pointer
back to the actual location of the .git folder inside the containing
project. To get at this directory, we need to run 'git rev-parse
--git-dir' instead of simply assuming that the local .git is the proper
directory.

Review: https://reviews.apache.org/r/41243