Forced horizontal scroll in docs code blocks.
Previously, we just wrapped the text and it was a mess. Now a horizontal
scroll bar will appear in code blocks instead of wrapping lines and
causing weirdness when trying to follow what's going on in the code.
It also makes it easier to copy and paste whole blocks of code.

Review: https://reviews.apache.org/r/41359