Added note in docs for installing wget, tar, git.
The 'Downloading Mesos' section of the getting started guide
assumes that you already have the 'wget', 'tar', and 'git' utilities
installed on your system before running through the 'System
Requirements' section below it. This commit adds a note to first run
through the 'System Requirments' section if problems are encountered
with downloading mesos via the provided instructions.

It also adds the 'wget', 'tar', and 'git' utilities to the list of
requirements for all systems in the 'System Requirements' section.
The two CentOS sections already had instructions for 'wget' and 'tar',
but now 'git' has been added to the list.

Review: https://reviews.apache.org/r/41362