Fixed getting started instructions for centos 7.1.
The current instructions are broken for installing the development tools
on centos 7.1 after updating the WANdisco SVN repo file. Specifically,
the updated svn installation requires libserf-1, which is not part of
the base yum repository.  We need to install the epel repository to get
access to libserf-1.

As part of this, I also update the instructions to install
java-1.8.0-openjdk-devel instead of java-1.7.0-openjdk-devel. This is
necessary because our instructions have us installing apache-maven
BEFORE installing java-1.7.0-openjdk-devel. As such, installing maven
installs the default jre from the base repository (java-1.8.0-openjdk)
as one of its dependencies. This causes problems with finding the proper
java.home in our mesos/configure script because of the mismatch between
the most up to date jre (1.8.0) and the most up to date development
tools (1.7.0). Alternative solutions to resolve this problem probably
exist (i.e. move our dependence on maven until AFTER our installation of
java-1.7.0-openjdk-devel), but this would likely break if running on a
system that already had the 1.8.0 jre installed for other purposes. Best
to just update to the newwest default version of the java development
tools.

Review: https://reviews.apache.org/r/41371