Added ubuntu dependence for libsasl2-modules.
This commit updates the Getting Started guide for ubuntu 14.04 to
explicitly list libsasl2-modules as a dependence for its system
requirements. It's possible that the stock ubuntu image comes with this
package automatically installed (e.g. the default ubuntu/trusty64 image
on atlas has it pre-installed). Regardless, we need to ensure that
it is installed before we build mesos so we list it here.

Review: https://reviews.apache.org/r/41383