Added flags to set size of completed task/framework history.
The default size of the buffers used to hold the state of completed
tasks/frameworks is very large. However, many users don't care much
about this information when requesting a master's state. Moreover, if a
large number of frameworks request this state simultaneously, the
master can quickly become overwhelmed because the process of generating
this state both blocks the master and takes up a lot of cycles. By
allowing the user to configure the size of the buffers used to hold
this state, we let the user decide how much state is needed.

This commit is based on a pull request generated by Felix Bechstein at:
https://github.com/apache/mesos/pull/82

Review: https://reviews.apache.org/r/42053/