Added unit test for framework/task history flags.
This commit adds tests to verify that the the max_frameworks and
max_tasks_per_frameworks flags for master work properly. Specifically,
we test to verify that the proper amount of history is maintained for
both 0 values to these flags as well as positive values <= to the total
number of frameworks and tasks per framework actually launched.

Review: https://reviews.apache.org/r/42212/