Added support file to autogenerate docs from endpoint help strings.
After building mesos from source, an executable called
mesos-endpoint-docs will be created in build/src. Invoke this executable
with the flag --dirname=../docs/endpoints to autogenerate documentation
for all endpoint strings and install it into the proper location.  In a
subsequent commit we will add a link to this directory in home.md.

Review: https://reviews.apache.org/r/42789