Added an example framework for consuming GPUs.
This framework is designed to show how to build a GPU capable
framework that can accept offers with GPUs and launch tasks that use
them. The key thing to remember is that the GPU_RESOURCES capability
must be set in `FrameworkInfo` in order for a framework to receive
resource offers from agents that contain GPUs.

Review: https://reviews.apache.org/r/48915