Reverted adding libelf as a default dynamic library in stout.
The libelf library is not a standard library installed on all Linux
systems, and it is not a library required to build stout in
general. It is only required for subsystems that happen to use the
<stout/elf.hpp> header. As such, we should only require this
dependency for those specific subsystems. If we ever decide to build
something into stout that depends on libelf by default, we can
revisit this issue then.

Review: https://reviews.apache.org/r/47610