Rearranged `expectedOffers` in the SlavesEndpointFullResources Test.
This change was required in response to the new refactoring for
resource enumeration. Since resources are now enumerated in a slightly
different order, the resulting JSON representing resources has a
slightly different ordering as well.

NOTE: We don't actually care about the ordering for this test. We only
care that the resulting resources match. For now we just fix the
ordering of resources in the JSON string, but we should consider
updating the test to compare the resulting `Resources` object
represented by the JSON, not the exact JSON strings themselves.

Review: https://reviews.apache.org/r/47708