Added ability to express dependencies between isolators.
Some isolators depend on other isolators. However, we previously did
not have a generic method of expressing these dependencies. We special
cased the `filesystem/*` isolators to make sure that dependencies on
them were satisfied, but no other dependencies could be expressed.

Now we use a vector to represent the pairing of isolator name to
isolator creator function, and the relative dependencies between each
isolator is implicit in the ordering of this vector. Previously, a
hashmap was used to hold this pairing, but this was inadequate because
hashmaps are inherently unordered. The new implementation using a
vector ensures everything is processed in the order it is listed.

Review: https://reviews.apache.org/r/48477