Updated elfio-3.1.patch to fix alignment bug properly.
The previous patch incorrectly fixed the alignment of an ELF note to 4
bytes (even for 64 bit ELF files). This happened to be OK for our
specific case of parsing '.note.ABI-tag' sections. However, the
alignment for notes in a 64 bit ELF section should be 8-byte aligned
(as the original code attempts to do). The problem with the original
code, however, is that it performs the alignment only on the name
string inside the note, rather than the entire note object itself.
This commit fixes this alignment issue.

Review: https://reviews.apache.org/r/49658