Added Nvidia devices as default devices to inject into every container.
This is a temporary hack to make Nvidia devices available to a
container when file system isolation is enabled. Ideally, we would
have a generic mechanism of injecting new devices into a continaer
from an isolator, but this feature is currently lacking so we need to
resort to hard-coding them for now.

In essence, this commit adds all Nvidia devices to the default list of
devices on machines that have Nvidia GPUs. Without the 'gpu/nvidia'
isolator enabled, this means that containers have free reign to use
these devices however they want. However, with the `'gpu/nvidia'
isolator enabled, they will have restricted access to only those GPUs
they have been allocated (though they will be able to see all of them
on the file system). These are similar semantics to the current
support for Nvidia GPUs without filesystem isolation.

In the future, we will restrict injecting these devices to only occur
when the 'nvidia/gpu' isolator is enabled.

Review: https://reviews.apache.org/r/49668