Updated mesos containerizer to checkpoint container runtime information.
This includes checkpointing both the container pid and the status of
the container upon exit. This also includes an update to tests to
account for new 'init' process semantics in a container. That is, the
name of the init process of the container is now "mesos-containerizer"
not "sh".

Review: https://reviews.apache.org/r/51407