Updated check for instantiating the "windows" launcher.
Previously, we allowed the posix launcher to specified on windows and
just blindly ignored it and instantiated a windows launcher anyway.
This was likely due to legacy code, not setting the default launcher
properly in flags.cpp. Now that we set the default to "windows"
properly, this hack is no longer needed.