Added special case for 'mnt' namespace when exec'ing into container.
We need to special case the 'mnt' namespace when exec'ing into a
container using the Mesos CLI in order to workaround MESOS-5727. To
find the proper `mnt` namespace to enter, We need to recursively look
at all children of the container's init-process and find the first
`mnt` namespace that doesn't match the mount namespace of the
init-process. If we find one, we enter it. If we don't find one, we
enter the init-process's `mnt` namespace.  In future versions of
mesos, the pod-executor will become the default executor and we will
always have a proper reference to the container whose `mnt` namespace
we should enter (thereby removing the need for this ugly hack).