Updated 'LinuxLauncher->fork()` with *enter* and *clone* namespaces.
Instead of only taking the namespaces to *clone* inside a newly forked
container, 'LinuxLauncher->fork()' now takes a list of namespaces to
enter inside a parent container before forking. When both an *enter*
and a *clone* for the same namespace are passed, we will first enter
the namespace of the parent, and then clone a new one.

Review: https://reviews.apache.org/r/53352