Updated containerizer->launch path to close IOSwitchboard FDs on error.
Previously, if the containizer launch path failed before actually
launching the container, the FDs allocated to the container by the
IOSwitchboard isolator would be leaked. This would lead to deadlock in
the destroy path because the IOSwitchboard does not shutdown until the
FDs it allocates to the container have been closed. Since the
switchboard doesn't shutdown, the future returned by its 'cleanup()'
function is never satisfied.

This commit makes sure to close the FDs under all failure cases in the
launch path.

Review: https://reviews.apache.org/r/56195