Added ujson to the list of modules as exceptions for pylint.
The extension-pkg-whitelist that 'ujson' is added to contains a list
of modules where C extensions may be loaded. Without being added to
this list, pylint would barf on importing the 'ujson' package.

In future commits we plan to use 'ujson' because it is significantly
faster than the standard 'json' library because of these C extensions.